src/main/java/au/com/easynebula/clienttransactionmanagement/producer/KafkaProducerConfig.java
package au.com.easynebula.clienttransactionmanagement.producer;

import org.apache.kafka.common.serialization.StringSerializer;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.context.annotation.Bean;
import org.springframework.context.annotation.Configuration;
import org.springframework.kafka.core.DefaultKafkaProducerFactory;
import org.springframework.kafka.core.KafkaTemplate;
import org.springframework.kafka.core.ProducerFactory;

import java.util.HashMap;
import java.util.Map;

import static org.apache.kafka.clients.producer.ProducerConfig.BOOTSTRAP_SERVERS_CONFIG;
import static org.apache.kafka.clients.producer.ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG;
import static org.apache.kafka.clients.producer.ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG;

@Configuration
public class KafkaProducerConfig {

	@Value("${kafka.server.url}")
	private String kafkaServerUrl;

	@Bean
	public ProducerFactory<String, String> producerFactory() {
		final Map<String, Object> configProps = new HashMap<>();
		configProps.put(BOOTSTRAP_SERVERS_CONFIG, kafkaServerUrl);
		configProps.put(KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
		configProps.put(VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
		return new DefaultKafkaProducerFactory<>(configProps);
	}

	@Bean
	public KafkaTemplate<String, String> kafkaTemplate() {
		return new KafkaTemplate<>(producerFactory());
	}
}
